General NBody inter- Ring and  Shiekh
Some painful but necessary changes are coming.  Restructuring to allow for generalized nbody interactions to be passed to the program in order to avoid lots of duplication as well as creating templates for real and complex routines.  This will simplify the code at the highest level and push all the ugly logic into contained classes.