Generaliing Code for Hamiltonian
Lot's of problems and debugging later the routines are working for the real hf cases.  There is a ways to go but significant progress has been done.  More testing to be done.  Also, I'm starting to think about a test suite.  Its in the infancy so likely it will change but there is something basic for now